feat: refactor radio interaction files to use 'rp' namespace and implement new handleAddRadio and handleRemoveRadio functions
