+++ src/interactions/buttons/index.ts
@@ -1,6 +1,6 @@
-import { changeRadio } from './hope/changeRadio';
-import { handleAddRadio } from './hope/handleAddRadio';
-import { handleRemoveRadio } from './hope/handleRemoveRadio';
+import { changeRadio } from './rp/changeRadio';
+import { handleAddRadio } from './rp/handleAddRadio';
+import { handleRemoveRadio } from './rp/handleRemoveRadio';
 import { handleMotusTry } from './motus/handleMotusTry';
 import { handleQuizButton } from './quiz/handleQuizButton';
 import { reportQuestionButton } from './quiz/reportQuestionButton';
